move private functions
