+++ nilearn/glm/_utils.py
@@ -2,147 +2,11 @@
 
 Authors: Bertrand Thirion, Matthew Brett, Ana Luisa Pinho, 2020
 """
-
-import csv
 from warnings import warn
 
 import numpy as np
-import pandas as pd
 import scipy.linalg as spl
 from scipy.stats import norm
-
-from nilearn._utils.helpers import stringify_path
-
-
-def _check_list_length_match(list_1, list_2, var_name_1, var_name_2):
-    """Check length match of two given lists to raise error if necessary."""
-    if len(list_1) != len(list_2):
-        raise ValueError(
-            "len(%s) %d does not match len(%s) %d"
-            % (str(var_name_1), len(list_1), str(var_name_2), len(list_2))
-        )
-
-
-def _read_events_table(table):
-    """Accept the path to en event.tsv file \
-    and loads it as a Pandas Dataframe.
-
-    Raises an error if loading fails.
-
-    Parameters
-    ----------
-    table : string
-        Accepts the path to an events file.
-
-    Returns
-    -------
-    loaded : pandas.Dataframe object
-        Pandas Dataframe with e events data.
-
-    """
-    try:
-        # kept for historical reasons, a lot of tests use csv with index column
-        loaded = pd.read_csv(table, index_col=0)
-    except:  # noqa: E722
-        raise ValueError(f"table path {table} could not be loaded")
-    if loaded.empty:
-        try:
-            loaded = pd.read_csv(table, sep="\t")
-        except:  # noqa: E722
-            raise ValueError(f"table path {table} could not be loaded")
-    return loaded
-
-
-def _check_and_load_tables(tables_, var_name):
-    """Check tables can be loaded in DataFrame to raise error if necessary."""
-    tables = []
-    for table_idx, table in enumerate(tables_):
-        table = stringify_path(table)
-        if isinstance(table, str):
-            loaded = _read_events_table(table)
-            tables.append(loaded)
-        elif isinstance(table, pd.DataFrame):
-            tables.append(table)
-        elif isinstance(table, np.ndarray):
-            pass
-        else:
-            raise TypeError(
-                f"{var_name} can only be a pandas DataFrames or a string. "
-                f"A {type(table)} was provided at idx {table_idx}"
-            )
-    return tables
-
-
-def _check_events_file_uses_tab_separators(events_files):
-    """Raise a ValueError if provided list of text based data files \
-    (.csv, .tsv, etc) do not enforce \
-    the :term:`BIDS` convention of using Tabs as separators.
-
-    Only scans their first row.
-    Does nothing if:
-        - If the separator used is :term:`BIDS` compliant.
-        - Paths are invalid.
-        - File(s) are not text files.
-
-    Does not flag comma-separated-values-files for compatibility reasons;
-    this may change in future as commas are not :term:`BIDS` compliant.
-
-    Parameters
-    ----------
-    events_files : str, List/Tuple[str]
-        A single file's path or a collection of filepaths.
-        Files are expected to be text files.
-        Non-text files will raise ValueError.
-
-    Returns
-    -------
-    None
-
-    Raises
-    ------
-    ValueError:
-        If value separators are not Tabs (or commas)
-
-    """
-    valid_separators = [",", "\t"]
-    if not isinstance(events_files, (list, tuple)):
-        events_files = [events_files]
-    for events_file_ in events_files:
-        if isinstance(events_file_, (pd.DataFrame)):
-            continue
-        try:
-            with open(events_file_) as events_file_obj:
-                events_file_sample = events_file_obj.readline()
-            """
-            The following errors are not being handled here,
-            as they are handled elsewhere in the calling code.
-            Handling them here will beak the calling code,
-            and refactoring that is not straightforward.
-            """
-        except OSError:  # if invalid filepath.
-            pass
-        else:
-            try:
-                csv.Sniffer().sniff(
-                    sample=events_file_sample,
-                    delimiters=valid_separators,
-                )
-            except csv.Error as e:
-                raise ValueError(
-                    "The values in the events file "
-                    "are not separated by tabs; "
-                    "please enforce BIDS conventions",
-                    events_file_,
-                ) from e
-
-
-def _check_run_tables(run_imgs, tables_, tables_name):
-    """Check fMRI runs and corresponding tables to raise error if necessary."""
-    if isinstance(tables_, (str, pd.DataFrame, np.ndarray)):
-        tables_ = [tables_]
-    _check_list_length_match(run_imgs, tables_, "run_imgs", tables_name)
-    tables_ = _check_and_load_tables(tables_, tables_name)
-    return tables_
 
 
 def z_score(pvalue, one_minus_pvalue=None):
